refactor: reorganize thrustcurves package; introduce MotorData type and API interaction functions
